[wip] move cb-based buildPerLocale
