+++ lib/construct/index.js
@@ -207,121 +207,13 @@ class Constructor extends EventEmitter {
     //
     spec.type = spec.type || 'webpack';
 
-    this.buildPerLocale({
+    this.builder.buildPerLocale({
       locale,
       progress,
       spec
     }, finish);
 
     return progress;
-  }
-
-  /**
-   * Downloads the package tarball based on the given `spec`, builds that `spec`
-   * given the written tarball and reports back via a progress stream
-   * @function buildPerLocale
-   * @param  {Object}   opts Options for the locale-specific build.
-   * @param    {String} opts.locale   BCP-47 locale name (e.g. en-US, fr, etc).
-   * @param    {Object} opts.spec     Specification object for the given build.
-   * @param    {Stream} opts.progress Progress "pseudo-stream" to report build progress on.
-   * @param  {Function} next Continuation to respond to when complete.
-   * @returns {undefined}
-   */
-  buildPerLocale(opts, next) {
-    const { progress, spec, locale } = opts;
-    const { app, topic } = this;
-    const id = uuid.v4();
-
-    //
-    // There are 3 events per ID. This is a stub of progress before we
-    // remove it in the next pass of the refactor as progress will need to
-    // exist in an external service. We use 2 here so that the `finished`
-    // event is the only 100 which gets sent when done is called
-    //
-    progress.start(id, 2, { locale });
-    const current = assign({
-      locale,
-      id
-    }, omit(spec, 'locales'));
-
-    app.contextLog.info('Start build for locale %s', locale, {
-      locale: locale,
-      name: spec.name,
-      version: spec.version,
-      env: spec.env,
-      promote: spec.promote,
-      id: id
-    });
-
-    /**
-      * Report the error to the developer before ending the async loop.
-      *
-      * @param {Error} err Failed step of the build process.
-      * @param {String} type Type of error
-      * @returns {undefined}
-      * @api private
-      */
-    const step = (err, type) => {
-      const key = this._key(spec);
-      if (err) {
-        progress.fail(err, id, { locale });
-        app.contextLog.error('Error in step %s for %s: %s', type, spec.name, err.message, {
-          locale: locale,
-          name: spec.name,
-          version: spec.version,
-          env: spec.env,
-          promote: spec.promote,
-          id: id
-        });
-
-        this.failures[key] = this.failures[key] || 0;
-        if (++this.failures[key] >= this.maxFailures)
-          return next(err);
-      }
-      next();
-    };
-
-    //
-    // Launch the build process with the specifications and attach
-    // a supervisor to communicate all events back to the developer.
-    //
-    progress.write({
-      locale,
-      progress: true,
-      message: `Queuing ${current.type} build for ${current.name}`,
-      id
-    });
-
-    const freshSpec = {
-      name: spec.name,
-      env: spec.env,
-      version: spec.version,
-      locale: locale,
-      type: spec.type,
-      promote: spec.promote
-    };
-
-    this.emit('queue', topic, freshSpec);
-    // MAKE ASYNC/AWAIT
-    return this.nsq.writer.publish(topic, freshSpec, (err) => {
-      if (err) {
-        app.contextLog.error('Build queue %s for %s env: %s failed %j', current.id, current.name, current.env);
-        return step(err);
-      }
-
-      this.emit('queued', topic, freshSpec);
-      app.contextLog.info('Finished queuing locale %s', locale, {
-        locale: locale,
-        env: spec.env,
-        version: spec.version,
-        name: spec.name,
-        id: id
-      });
-
-      progress.done(id, { locale });
-      return void step();
-    });
-
   }
 
   /**
